pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-generative-models"
version = "0.1.0"
description = "Coursework scaffolding for Deep Generative Models."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Fatemeh Raeijian"}
]
dependencies = [
  "torch",
  "torchvision",
  "numpy",
  "matplotlib",
  "tensorboard",
  "scikit-learn",
  "notebook",
  "ipywidgets",
  "pyyaml",
]

[project.optional-dependencies]
dev = [
  "black",
  "ruff",
  "pytest",
]

[project.urls]
Homepage = "https://example.com"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

